requirements.txt
libclang~=16.0.0
pefile~=2023.2.7
pyxbe~=1.0.2
